project/db/src/requirements.txt
Flask==1.0.2
Flask-Login==0.4.1
Flask-SQLAlchemy==2.3.2
python-dateutil==2.8.0
SQLAlchemy==1.3.1
itsdangerous==2.0.1
jinja2==3.0.3
Werkzeug~=2.0.0